feat(nav): Separate org and user dropdowns
