+++ static/app/views/nav/orgDropdown.spec.tsx
@@ -1,7 +1,7 @@
 import {OrganizationFixture} from 'sentry-fixture/organization';
 import {UserFixture} from 'sentry-fixture/user';
 
-import {render, screen, userEvent, waitFor} from 'sentry-test/reactTestingLibrary';
+import {render, screen, userEvent} from 'sentry-test/reactTestingLibrary';
 
 import ConfigStore from 'sentry/stores/configStore';
 import OrganizationsStore from 'sentry/stores/organizationsStore';
@@ -36,59 +36,6 @@ describe('OrgDropdown', function () {
       'href',
       `/organizations/${organization.slug}/settings/teams/`
     );
-  });
-
-  it('displays user info and links', async function () {
-    render(<OrgDropdown />, {organization});
-
-    await userEvent.click(screen.getByRole('button', {name: 'Toggle organization menu'}));
-
-    expect(screen.getByText('Foo Bar')).toBeInTheDocument();
-    expect(screen.getByText('foo@example.com')).toBeInTheDocument();
-
-    expect(screen.getByRole('link', {name: 'User Settings'})).toHaveAttribute(
-      'href',
-      '/settings/account/'
-    );
-    expect(screen.getByRole('link', {name: 'User Auth Tokens'})).toHaveAttribute(
-      'href',
-      '/settings/account/api/'
-    );
-  });
-
-  it('can sign out', async function () {
-    const mockLogout = MockApiClient.addMockResponse({
-      url: '/auth/',
-      method: 'DELETE',
-    });
-
-    render(<OrgDropdown />, {organization});
-
-    await userEvent.click(screen.getByRole('button', {name: 'Toggle organization menu'}));
-
-    await userEvent.click(screen.getByText('Sign Out'));
-
-    await waitFor(() => {
-      expect(mockLogout).toHaveBeenCalled();
-    });
-  });
-
-  it('hides admin link if user is not admin', async function () {
-    render(<OrgDropdown />, {organization});
-
-    await userEvent.click(screen.getByRole('button', {name: 'Toggle organization menu'}));
-
-    expect(screen.queryByRole('link', {name: 'Admin'})).not.toBeInTheDocument();
-  });
-
-  it('shows admin link if user is admin', async function () {
-    ConfigStore.set('user', UserFixture({isSuperuser: true}));
-
-    render(<OrgDropdown />, {organization});
-
-    await userEvent.click(screen.getByRole('button', {name: 'Toggle organization menu'}));
-
-    expect(screen.getByRole('link', {name: 'Admin'})).toHaveAttribute('href', `/manage/`);
   });
 
   it('can switch orgs', async function () {
