+++ static/app/views/nav/orgDropdown.tsx
@@ -2,12 +2,10 @@ import {useTheme} from '@emotion/react';
 import styled from '@emotion/styled';
 import orderBy from 'lodash/orderBy';
 
-import {logout} from 'sentry/actionCreators/account';
 import {OrganizationAvatar} from 'sentry/components/core/avatar/organizationAvatar';
 import {Button} from 'sentry/components/core/button';
 import {DropdownMenu, type MenuItemProps} from 'sentry/components/dropdownMenu';
 import OrganizationBadge from 'sentry/components/idBadge/organizationBadge';
-import UserBadge from 'sentry/components/idBadge/userBadge';
 import {IconAdd} from 'sentry/icons';
 import {t, tn} from 'sentry/locale';
 import ConfigStore from 'sentry/stores/configStore';
@@ -15,12 +13,11 @@ import OrganizationsStore from 'sentry/stores/organizationsStore';
 import {useLegacyStore} from 'sentry/stores/useLegacyStore';
 import {isDemoModeActive} from 'sentry/utils/demoMode';
 import {localizeDomain, resolveRoute} from 'sentry/utils/resolveRoute';
-import useApi from 'sentry/utils/useApi';
 import useOrganization from 'sentry/utils/useOrganization';
 import useProjects from 'sentry/utils/useProjects';
-import {useUser} from 'sentry/utils/useUser';
 import {useNavContext} from 'sentry/views/nav/context';
 import {NavLayout} from 'sentry/views/nav/types';
+import {makeProjectsPathname} from 'sentry/views/projects/pathname';
 
 function createOrganizationMenuItem(): MenuItemProps {
   const configFeatures = ConfigStore.get('features');
@@ -54,12 +51,10 @@ export function OrgDropdown({
   className?: string;
   hideOrgLinks?: boolean;
 }) {
-  const api = useApi();
   const theme = useTheme();
 
   const config = useLegacyStore(ConfigStore);
   const organization = useOrganization();
-  const user = useUser();
 
   // It's possible we do not have an org in context (e.g. RouteNotFound)
   // Otherwise, we should have the full org
@@ -74,10 +69,6 @@ export function OrgDropdown({
 
   const {layout} = useNavContext();
   const isMobile = layout === NavLayout.MOBILE;
-
-  function handleLogout() {
-    logout(api);
-  }
 
   return (
     <DropdownMenu
@@ -96,6 +87,7 @@ export function OrgDropdown({
           />
         </OrgDropdownTrigger>
       )}
+      position="right-start"
       minMenuWidth={200}
       items={[
         {
@@ -115,6 +107,12 @@ export function OrgDropdown({
               label: t('Organization Settings'),
               to: `/organizations/${organization.slug}/settings/`,
               hidden: !hasOrgRead || hideOrgLinks,
+            },
+            {
+              key: 'projects',
+              label: t('Projects'),
+              to: makeProjectsPathname({path: '/', orgSlug: organization.slug}),
+              hidden: hideOrgLinks,
             },
             {
               key: 'members',
@@ -158,38 +156,6 @@ export function OrgDropdown({
                 }),
                 createOrganizationMenuItem(),
               ],
-            },
-          ],
-        },
-        {
-          key: 'user',
-          label: (
-            <SectionTitleWrapper>
-              <UserBadge user={user} avatarSize={32} />
-            </SectionTitleWrapper>
-          ),
-          textValue: t('User Summary'),
-          children: [
-            {
-              key: 'user-settings',
-              label: t('User Settings'),
-              to: '/settings/account/',
-            },
-            {
-              key: 'user-auth-tokens',
-              label: t('User Auth Tokens'),
-              to: '/settings/account/api/',
-            },
-            {
-              key: 'admin',
-              label: t('Admin'),
-              to: '/manage/',
-              hidden: !user?.isSuperuser,
-            },
-            {
-              key: 'signout',
-              label: t('Sign Out'),
-              onAction: handleLogout,
             },
           ],
         },
