+++ static/app/views/nav/primary/index.tsx
@@ -29,6 +29,7 @@ import {PrimaryNavigationServiceIncidents} from 'sentry/views/nav/primary/servic
 import {PrimaryNavigationWhatsNew} from 'sentry/views/nav/primary/whatsNew';
 import {NavTourElement, StackedNavigationTour} from 'sentry/views/nav/tour/tour';
 import {NavLayout, PrimaryNavGroup} from 'sentry/views/nav/types';
+import {UserDropdown} from 'sentry/views/nav/userDropdown';
 
 function SidebarBody({children}: {children: React.ReactNode}) {
   const {layout} = useNavContext();
@@ -151,9 +152,6 @@ export function PrimaryNavigationItems() {
       <SidebarFooter>
         <ChonkOptInBanner collapsed="never" />
         <PrimaryNavigationHelp />
-
-        <SeparatorItem />
-
         <PrimaryNavigationWhatsNew />
         <Hook
           name="sidebar:try-business"
@@ -165,6 +163,8 @@ export function PrimaryNavigationItems() {
         <ErrorBoundary customComponent={null}>
           <PrimaryNavigationOnboarding />
         </ErrorBoundary>
+        <SeparatorItem hasMargin />
+        <UserDropdown />
       </SidebarFooter>
     </Fragment>
   );
